Fixed new_attribute_error and added more tests for property
